Add HD6301Y0 data sheet and document ideas for dumping the ROM
In this commit my changes and additions to Markdown files use a text format
more optimized for readability as a text file (reference-style links,
wrapping before column 80, more prominent level 2 headers, etc.). If this
isn't desired, feel free to change it (either by amending this commit or
reformatting in a subsequent commit) and I will follow the original
formatting style.